refactor(CrowdAdd.vue；CreatePolicyWithLabelSquare.vue; ModelLabelIndex.vue): 修复【一键投放】中不选标签时，渲染人群规则时报错的问题；解决标签广场 - 模型标签列表搜索的问题；标签广场 - 模型标签删除时添加确认弹窗；一键投放 - 模型标签样式修改

+++ src/views/LabelSquare/ModelLabel/ModelLabelIndex.vue
@@ -10,7 +10,7 @@
         <el-input
           placeholder="支持按标签名、ID搜索"
           class="header-input"
-          v-model="filter.tagName"
+          v-model="filter.condition"
           @keyup.enter.native="fetchData"></el-input>
         <i class="el-icon-cc-search icon-fixed" @click="fetchData"></i>
       </div>
@@ -130,8 +130,14 @@ export default {
   methods: {
     // 删除
     handleDelete (id) {
-      this.$service.deleteModelTag(id, '删除成功').then(() => {
-        this.fetchData()
+      this.$confirm('确定要删除吗?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        this.$service.deleteModelTag(id, '删除成功').then(() => {
+          this.fetchData()
+        })
       })
     },
     handleEdit (row) {
